Extract MessageComposerFragment and VoiceRecorderFragment from TimelineFragment
